third_party/blink/web_tests/crypto/subtle/aes-cbc/wrap-rsa-non-extractable-expected.txt

Test wrapping an RSA key with AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a key to wrap...
PASS key.algorithm.name is 'RSASSA-PKCS1-v1_5'
Importing a key encryption key...
PASS wrappingKey.algorithm.name is 'AES-CBC'
wrapping the key...
PASS Rejected: InvalidAccessError: key is not extractable
PASS successfullyParsed is true

TEST COMPLETE
